+++ pkg/dto/experiment_response.go
@@ -4,6 +4,7 @@ import "time"
 
 type ExperimentResponse struct {
 	ID                   uint       `json:"id"`
+	UserID               uint       `json:"user_id"`
 	Title                string     `json:"title"`
 	Comments             string     `json:"comments"`
 	MeasurementStartTime *time.Time `json:"measurement_start_time,omitempty"`